refactor: split storage_collaborative.py into focused modules
- Created storage_base.py (35 lines): BaseStorageResource with common functionality
- Created storage_bucket.py (148 lines): BucketListResource for file listing
- Created storage_locks.py (281 lines): FileLockResource and FileUnlockResource
- Refactored storage_collaborative.py (621 lines, was 1010): Kept FileCopyResource and FileInfoResource
- Fixed pylint too-many-lines warning (1010 → 621 lines, -38%)
- Maintained backward compatibility via re-exports
- Fixed import order and removed unused imports
- All 163 tests passing
